Create unified Navbar component
