+++ localization/messages/hr.json
@@ -55,11 +55,11 @@
   "nav.home": "Početna",
   "nav.myLessons": "Moje instrukcije",
   "nav.myRequests": "Moji zahtjevi",
-  "nav.switch": "Prebaci se na instruktora",
-  "nav.switch.tutor": "Prebaci se na studenta",
   "nav.publicRequests": "Javni zahtjevi",
   "nav.tutorResponses": "Moji odgovori",
-  "nav.becomeTutor": "Postani instruktor",
+  "nav.becomeATutor": "Postani instruktor",
+  "nav.switchToTutor": "Prebaci se na instruktora",
+  "nav.switchToStudent": "Prebaci se na studenta",
 
   "home.rewards": "Nagrade",
   "home.lessons": "Lekcije",
