+++ pages/student/new-request.tsx
@@ -2,7 +2,7 @@ import type { NextPage } from 'next';
 import Head from 'next/head';
 import { useRouter } from 'next/router';
 
-import { StudentsNavbar, NewRequestForm } from '@modules';
+import { NewRequestForm, Navbar } from '@modules';
 
 const RequestsPage: NextPage = () => {
   const router = useRouter();
@@ -12,7 +12,8 @@ const RequestsPage: NextPage = () => {
         <title>InstruGo | New request</title>
       </Head>
 
-      <StudentsNavbar />
+      <Navbar />
+
       <NewRequestForm onFinish={() => router.push('/student/requests')} />
     </div>
   );
