+++ src/modules/index.ts
@@ -2,13 +2,11 @@ export * from './RegistrationForm/RegistrationForm';
 export * from './LoginForm/LoginForm';
 export * from './SimpleHeader/Simpleheader';
 export * from './NewRequestForm/NewRequestForm';
-export * from './StudentsNavbar/StudentsNavbar';
 export * from './Rewards/Rewards';
 export * from './NewRequestButton/NewRequestButton';
 export * from './RequestDetails/RequestDetails';
 export * from './PublicRequestsContainer/PublicRequestsContainer';
 export * from './PublicRequestDetails/PublicRequestDetails';
-export * from './TutorsNavbar/TutorsNavbar';
 export * from './PrivateProfile/PrivateProfile';
 export * from './LessonDetails/LessonDetails';
 export * from './LessonDetails/LessonDetailsAfterStudent';
@@ -22,3 +20,4 @@ export * from './LessonCard/LessonCard';
 export * from './LessonsContainer/LessonsContainer';
 export * from './LessonsWithFilter/LessonsWithFilter';
 export * from './LessonsTable/LessonsTable';
+export * from './Navbar/Navbar';
